# 05/results/swaggerDocumentation.json
{"openapi":"3.0.1","info":{"title":"Rest Application","description":"Rest Application","termsOfService":"http://swagger.io/terms/","license":{"name":"Apache 2.0","url":"http://springdoc.org"},"version":"1.0"},"servers":[{"url":"http://localhost:8080/","description":"Generated server url"}],"paths":{"/tour/":{"get":{"tags":["tour-controller"],"summary":"Get all tours","operationId":"getTours","responses":{"200":{"description":"OK","content":{"application/json":{}}}}},"put":{"tags":["tour-controller"],"summary":"Update tour","operationId":"updateTour","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Tour"}}},"required":true},"responses":{"200":{"description":"Updated the tour","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Tour"}}}},"404":{"description":"Invalid tour"}}},"post":{"tags":["tour-controller"],"summary":"Add new tour","operationId":"createTour","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Tour"}}},"required":true},"responses":{"201":{"description":"New tour added","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Tour"}}}}}}},"/location/":{"get":{"tags":["location-controller"],"summary":"Get all locations","operationId":"getLocations","parameters":[{"name":"countryId","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{}}}}},"put":{"tags":["location-controller"],"summary":"Update location","operationId":"updateLocation","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Location"}}},"required":true},"responses":{"200":{"description":"Updated the location","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Location"}}}},"404":{"description":"Invalid location"}}},"post":{"tags":["location-controller"],"summary":"Add new location","operationId":"createLocation","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Location"}}},"required":true},"responses":{"201":{"description":"New location added","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Location"}}}}}}},"/customer/":{"get":{"tags":["customer-controller"],"summary":"Get all customers","operationId":"getCustomers","responses":{"200":{"description":"OK","content":{"application/json":{}}}}},"put":{"tags":["customer-controller"],"summary":"Update customer","operationId":"updateCustomer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Customer"}}},"required":true},"responses":{"404":{"description":"Invalid customer"},"200":{"description":"Updated the customer","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Customer"}}}}}},"post":{"tags":["customer-controller"],"summary":"Add new customer","operationId":"createCustomer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Customer"}}},"required":true},"responses":{"201":{"description":"New customer added","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Customer"}}}}}}},"/country/":{"get":{"tags":["country-controller"],"summary":"Get all countries","operationId":"getCountries","responses":{"200":{"description":"OK","content":{"application/json":{}}}}},"put":{"tags":["country-controller"],"summary":"Update country","operationId":"updateLocation_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Country"}}},"required":true},"responses":{"200":{"description":"Updated the country","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Country"}}}},"404":{"description":"Invalid country"}}},"post":{"tags":["country-controller"],"summary":"Add new country","operationId":"createCountry","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Country"}}},"required":true},"responses":{"201":{"description":"New country added","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Country"}}}}}}},"/tour/{id}":{"get":{"tags":["tour-controller"],"summary":"Get tour by id","operationId":"getTour","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"Invalid tour id"},"200":{"description":"Found the tour","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Tour"}}}}}},"delete":{"tags":["tour-controller"],"summary":"Delete tour","operationId":"deleteTour","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Deleted the tour","content":{"application/json":{}}},"404":{"description":"Invalid tour"}}}},"/location/{id}":{"get":{"tags":["location-controller"],"summary":"Get location by id","operationId":"getLocation","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"Invalid location id"},"200":{"description":"Found the location","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Location"}}}}}},"delete":{"tags":["location-controller"],"summary":"Delete location","operationId":"deleteLocation","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"Invalid location"},"200":{"description":"Deleted the location","content":{"application/json":{}}}}}},"/customer/{id}":{"get":{"tags":["customer-controller"],"summary":"Get customer by id","operationId":"getCustomer","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Found the customer","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Customer"}}}},"404":{"description":"Invalid customer id"}}},"delete":{"tags":["customer-controller"],"summary":"Delete customer","operationId":"deleteCustomer","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Deleted the customer","content":{"application/json":{}}},"404":{"description":"Invalid customer"}}}},"/country/{id}":{"get":{"tags":["country-controller"],"summary":"Get country by id","operationId":"getCountry","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"Invalid country id"},"200":{"description":"Found the country","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Country"}}}}}},"delete":{"tags":["country-controller"],"summary":"Delete country","operationId":"deleteCountry","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Deleted the country","content":{"application/json":{}}},"404":{"description":"Invalid country"}}}}},"components":{"schemas":{"Country":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"links":{"type":"array","items":{"$ref":"#/components/schemas/Link"}}}},"Link":{"type":"object","properties":{"rel":{"type":"string"},"href":{"type":"string"},"hreflang":{"type":"string"},"media":{"type":"string"},"title":{"type":"string"},"type":{"type":"string"},"deprecation":{"type":"string"},"profile":{"type":"string"},"name":{"type":"string"}}},"Location":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"country":{"$ref":"#/components/schemas/Country"},"links":{"type":"array","items":{"$ref":"#/components/schemas/Link"}}}},"Tour":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"location":{"$ref":"#/components/schemas/Location"},"links":{"type":"array","items":{"$ref":"#/components/schemas/Link"}}}},"Customer":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"tours":{"type":"array","items":{"$ref":"#/components/schemas/Tour"}},"links":{"type":"array","items":{"$ref":"#/components/schemas/Link"}}}}}}}